List all open margin positions
Get a list of currently held open margin positions. Pre-req for a
command to watch your margin position and close it out at a certain
threshold. I can't find a way to do this in the UI and don't like
watching my position as much as would be needed to close out at
a certain point.